+++ src/test/java/io/beanmapper/testmodel/collections/CollTargetMapNotAnnotated.java
@@ -1,0 +1,10 @@
+package io.beanmapper.testmodel.collections;
+
+import java.util.Map;
+import java.util.TreeMap;
+
+public class CollTargetMapNotAnnotated {
+
+    public Map<String, Long> map = new TreeMap<>();
+
+}